requirements.txt
streamlit==1.32.0
PyPDF2==3.0.0
transformers==4.40.0
torch==2.2.0
python-dotenv==1.0.0
soundfile==0.12.1
librosa==0.10.1
openai-whisper==20231117